requirements.txt
requests
mysql-connector-python==8.3.0